requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
pymysql==1.1.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.1
scikit-learn==1.2.2
gunicorn==20.1.0
python-dotenv==1.0.0
